Use cookies to track auth. Use include list for redirects.
